refactor(core): move OTLP exporter into tracing folder
